+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to Gogs are documented in this file.
 
 - Submodules using `ssh://` protocol and a port number are not rendered correctly. [#4941](https://github.com/gogs/gogs/issues/4941)
 - Missing link to user profile on the first commit in commits history page. [#7404](https://github.com/gogs/gogs/issues/7404)
+- Unable to override email templates in custom directory. [#7905](https://github.com/gogs/gogs/pull/7905)
 
 ## 0.13.2